Create a GitHub release with the built binaries
Rename the 'Build' job to 'Compile' and add a 'Release' job which runs
after both of the Compile jobs have succeded.  This new job creates a
GitHub release and attaches the binaries if it is the first successful
build of the particular release from the master branch.

Also enable compiling (but not releasing) submitted pull requests.