Work around special chars 'with' parameter issue
Something in the graphql-action or action runner seems to have started
treating the query parameter value as plain YAML recently even though it
is a literal block.  This causes the workflow to fail.

Work around this issue by removing any spaces after colons in the query
parameter values as this makes them valid plain YAML strings as well as
remaining valid GraphQL.