Build release specified in file instead of latest
Instead of always building the latest version of Dropbear, store the
version to build in a new file: 'dropbear-version'.  This means that the
source for a particular release tag will build that release rather than
always building the latest.  It also means that version updates (and any
necessary changes or patches) can be managed via Pull Requests instead
of a scheduled build of master.