Build & deploy JAR to allow easy use as dependency
Combine all of the built binaries into a single JAR in the relevant ABI
folders under lib (renamed to the pattern lib*.so) so that this can be
called up as a dependency to an Android Studio project.  Also copy the
Dropbear license to META-INF/LICENSE so it is kept with the binaries.

Upload the built JAR to the GitHub package registry.