Use HOME env var when looking for id_dropbear
dbclient uses the value of HOME by default when looking for
~/.ssh/known_hosts, falling back to getpwuid()->pw_dir if HOME is not
set.  This is very useful under Android as getpwuid()->pw_dir is
hardcoded to "/data" for the app users but HOME defaults to the app's
data directory and can also be overridden when running a process.

Unfortunately when locating ~/.ssh/id_dropbear only the value from
getpwuid() is used, preventing use of a default identity.  Resolve the
issue by patching the Dropbear source to use the value of HOME if set
(submitted upstream as mkj/dropbear#137).