Actions: Set ANDROID_NDK_HOME to oldest installed
As part of updates to the GitHub Actions runners, ANDROID_NDK_HOME is
no-longer set.  Use info from sdkmanager to build this ourselves.

Use the oldest installed NDK to build as GitHub's policy is to have the
two latest LTS versions installed.  This means that the oldest installed
won't be ancient but may well support a lower minimum API level.