Fix build/checks trigger from version update PRs
The branches-ignore entries for update-version added in 99027d8bcf fixed
the issue of both pull_request and push triggering separate workflow
runs when a version update PR has additional commits pushed.  However,
it also stopped the version update PRs triggering the workflows at all
when created as actions by the actions user don't trigger further
workflow runs (but the push used to trigger a workflow as it runs as my
user via a deploy key).

Fix this regression by removing the branch from branches-ignore and
adding conditions to skip the push event when a user who isn't actions
pushes to the update-version branch.