Mention minimum API level in release notes
As the PLATFORM variable in the build script defines the minimum API
level that the binaries will work on, write this to a file in the
buildinfo dir and then reference this to insert the value into the body
of the created release.

Also stop reading the ndkver file into an environment variable, just
read the file when the value is needed.