Backport -D to specify server authorized_keys path
It is not straight-forward to create the directory /.ssh on more recent
Android devices (which ends up being necessary due to bionic being
hard-coded to return / for home directory paths).

Happily, a new '-D' option for overriding the default path for
authorized_keys was recently added to upstream so backport this until it
is included in a release.

Resolves #61.